// src/types/header-category.ts
import type { RefObject } from 'react';
import type { CategoryItem } from '@/types/types.ts';

export type HeaderCategoryNavItemType = 'default' | 'external' | 'popover';

export type HeaderCategoryNavItemData = CategoryItem & {
  type: HeaderCategoryNavItemType;
  options?: {
    label: string;
    routing: string;
  }[];
};

export type HeaderCategoryOverflowState = {
  navRef: RefObject<HTMLElement | null>;
  hasOverflow: boolean;
  isScrollEnd: boolean;
};